!250 修复entity中Lock注解判断异常

+++ src/main/java/cn/cerc/mis/ado/EntityMany.java
@@ -83,16 +83,13 @@ public class EntityMany<T extends EntityImpl> extends EntityHome<T> implements I
     }
 
     public void deleteAll() {
-        boolean check = EntityHelper.create(clazz).lockedField().isPresent();
+        var field = EntityHelper.create(clazz).lockedField();
         query.setReadonly(false);
         try {
             query.first();
             while (!query.eof()) {
-                if (check) {
-                    var entity = query.current().asEntity(clazz);
-                    if (entity.isLocked())
-                        throw new RuntimeException("record is locked");
-                }
+                if (field.isPresent() && query.getBoolean(field.get().getName()))
+                    throw new RuntimeException("record is locked");
                 query.delete();
             }
         } finally {
@@ -103,11 +100,11 @@ public class EntityMany<T extends EntityImpl> extends EntityHome<T> implements I
     public void deleteAll(List<T> list) {
         query.setReadonly(false);
         try {
-            boolean check = EntityHelper.create(clazz).lockedField().isPresent();
+            var field = EntityHelper.create(clazz).lockedField();
             for (T entity : list) {
                 if (entity.findRecNo() < 0)
                     throw new RuntimeException("delete fail, entity not in query");
-                if (check && entity.isLocked())
+                if (field.isPresent() && query.getBoolean(field.get().getName()))
                     throw new RuntimeException("record is locked");
                 query.delete();
             }
